finished the practice project
All items are centered on the page (horizontally, not vertically).
The title is centered on the page.
There is 32px between the title and the 'items.'
There is 52px between each item.
The items are arranged horizontally on the page.
The items are only 200px wide and the text wraps.
The item text is centered.